requirements.txt
asgiref==3.4.1
Django==4.0
django-taggit==3.0.0
sqlparse==0.4.2
tzdata==2021.5
